--- app/services/solidus_bolt/webhooks/sorter.rb
@@ -1,37 +1,0 @@
-# frozen_string_literal: true
-
-module SolidusBolt
-  module Webhooks
-    class Sorter
-      attr_reader :params
-
-      def self.call(params)
-        new(params).call
-      end
-
-      def initialize(params)
-        @params = params
-      end
-
-      def call
-        handler&.call(params)
-      end
-
-      private
-
-      def handler
-        class_name = event_type.split(".")
-                               .map(&:capitalize)
-                               .join
-
-        return unless SolidusBolt::Webhooks::Handlers.const_defined?(class_name)
-
-        SolidusBolt::Webhooks::Handlers.const_get(class_name)
-      end
-
-      def event_type
-        params[:type]
-      end
-    end
-  end
-end
