+++ app/webhooks/solidus_bolt/handlers/base_handler.rb
@@ -1,0 +1,21 @@
+# frozen_string_literal: true
+
+module SolidusBolt
+  module Handlers
+    class BaseHandler
+      attr_reader :params
+
+      def initialize(params)
+        @params = params
+      end
+
+      def self.call(params)
+        new(params).call
+      end
+
+      def call
+        raise NotImplementedError, 'Missing #call method on class'
+      end
+    end
+  end
+end
